Updated build, test, and workflow. Included a fully operational workflow (uncomment to use)

+++ build/rollup.test.node.config.js
@@ -25,7 +25,10 @@ if (build.isUnnecessaryBuild(input, output)) {
 }
 
 // 1. Clean prior builds
-fs.rmdirSync(rootdir, { recursive: true })
+if (fs.existsSync(rootdir)) {
+  console.log('Cleaning directory:', rootdir)
+  fs.rmdirSync(rootdir, { recursive: true })
+}
 
 // Identify plugins
 const plugins = [
